[mac_ai] visualizations: llm_load_test: finish remoting llm-load-test from the LTS
